migrate non pages components: /projects
